requirements.txt
colorama==0.4.3
dnspython==1.16.0
loguru==0.4.1
mysql-connector-python==8.0.19
numpy==1.18.3
pandas==1.0.3
protobuf==3.6.1
python-dateutil==2.8.1
pytz==2020.1
six==1.14.0
win32-setctime==1.0.1
